changed folder structure to something that makes sense
