+++ backend/services/analyzePatterns.js
@@ -1,0 +1,7 @@
+function  analyzePatterns (req, res, db, apiKey, model) {
+
+     return res.json({
+      message: "Hello from analysePatterns"
+    });
+}
+module.exports = { analyzePatterns };
