+++ backend/services/promptBuilder.js
